backend/requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.3
djangorestframework==3.14.0
exceptiongroup==1.0.4
flake8==5.0.4
iniconfig==1.1.1
Markdown==3.4.1
mccabe==0.7.0
packaging==21.3
Pillow==9.3.0
pluggy==1.0.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.2.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
